+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
-- HtmlFilter: <script> content should be filtered out.
+### Fixed
 
+- HtmlFilter: <script> content should be filtered out.
+- HtmlFilter: only for "keywords" and "description" meta tags "content" attr should be treated as
+  string.
 
 ## 1.7.1 - 2017-05-01